Update detector.c
Output of file now will take the input image (dog.jpg) and append
'_predicitions' to the file name, so that the resulting image's filename
has relevance to the input image (dog_predictions.jpg).

Added three support functions to facilitate this feature.
   find_char
   extract_str
   string_cat

find_char linearly searches for a character to match, if it does not
match, the function will return -1

extract_str will return a sub-string from indices begin to end and store
it into dst

string_cat will append the source (src) to the destination (dst). It
assumes that dst has enough allocated memory to support the
concatenation